Move more notification queries to api/notifications
