+++ client/src/components/Notifications/NotificationActions.test.ts
@@ -5,7 +5,7 @@ import flushPromises from "flush-promises";
 import { setActivePinia } from "pinia";
 import Vue from "vue";
 
-import type { UserNotification } from "@/components/Notifications/index";
+import type { UserNotification } from "@/api/notifications";
 import { generateMessageNotification, generateNewSharedItemNotification } from "@/components/Notifications/test-utils";
 import { useNotificationsStore } from "@/stores/notificationsStore";
