+++ client/src/components/Notifications/test-utils.ts
@@ -1,8 +1,10 @@
-import type { components } from "@/api/schema";
-import type { MessageNotification, SharedItemNotification, UserNotification } from "@/components/Notifications/index";
-
-type NotificationVariants = components["schemas"]["NotificationVariant"];
-type NewSharedItemNotificationContentItemType = components["schemas"]["NewSharedItemNotificationContent"]["item_type"];
+import type {
+    MessageNotification,
+    NewSharedItemNotificationContentItemType,
+    NotificationVariants,
+    SharedItemNotification,
+    UserNotification,
+} from "@/api/notifications";
 
 export function generateRandomItemType() {
     const itemTypes = ["history", "workflow", "visualization", "page"];
